+++ CS559_project2v2/Starfield.h
@@ -17,7 +17,7 @@ class Starfield : public Object
 public:
 	Starfield(void);
 	void TakeDown();
-	bool Initialize(double inner_radius, double depth, int stars);
+	bool Initialize(float inner_radius, float depth, int stars);
 	virtual void Draw(const glm::mat4 & projection, glm::mat4 modelview, const glm::ivec2 & size, const float time = 0);
 
 	Shader solid_color;
